Refactored HW3
